Refactoring folders.
