+++ src/server/domain/store/snapshot.go
