+++ src/server/domain/vm/aggregate/snapshot.go
