+++ src/server/domain/vm/entity.go
