+++ src/server/domain/vm/player/projector.go
@@ -1,0 +1,12 @@
+package player
+
+import "github.com/domain-query-language/dql-server/src/server/domain/vm"
+
+type Projector interface {
+
+	Reset()
+
+	Apply(event vm.Event) error
+
+	Version() int
+}
