+++ src/server/infrastructure/domain/store/stream_bolt.go
