--- src/server/vm/aggregate/stream.go
@@ -1,12 +1,0 @@
-package aggregate
-
-import "github.com/domain-query-language/dql-server/src/server/vm"
-
-type Stream interface {
-
-	Reset()
-
-	Seek(version int)
-
-	Next() vm.Event
-}
