--- src/server/vm/handler/query/query.go
@@ -1,8 +1,0 @@
-package query
-
-import "github.com/domain-query-language/dql-server/src/server/vm"
-
-type Query interface {
-
-	Id() vm.Identifier
-}
