--- src/server/vm/handler/query/result.go
@@ -1,6 +1,0 @@
-package query
-
-type Result interface {
-
-	MarshalJSON() ([]byte, error)
-}
